bugFix(arm): Fixed armID

+++ src/main/java/frc/robot/subsystems/arm/ArmConstants.java
@@ -13,7 +13,7 @@ public static final boolean kBrake = false;
 public static final Angle kLevelFourAngle = Degrees.of(0);
 public static final Angle kLevelThreeAngle = Degrees.of(0);
 public static final Angle kLevelTwoAngle = Degrees.of(0);
-public static final double armID = 15;
+public static final int armID = 15;
 
 
 }
